@@ -13,52 +13,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.5",
-    "@angular/common": "8.2.5",
-    "@angular/compiler": "8.2.5",
-    "@angular/core": "8.2.5",
-    "@angular/fire": "^5.2.1",
-    "@angular/forms": "8.2.5",
-    "@angular/platform-browser": "8.2.5",
-    "@angular/platform-browser-dynamic": "8.2.5",
-    "@angular/router": "8.2.5",
+    "@angular/animations": "11.2.4",
+    "@angular/common": "11.2.4",
+    "@angular/compiler": "11.2.4",
+    "@angular/core": "11.2.4",
+    "@angular/fire": "^6.1.4",
+    "@angular/forms": "11.2.4",
+    "@angular/platform-browser": "11.2.4",
+    "@angular/platform-browser-dynamic": "11.2.4",
+    "@angular/router": "11.2.4",
     "@ng-bootstrap/ng-bootstrap": "^5.1.1",
     "bootstrap": "^4.3.1",
     "core-js": "^3.2.1",
-    "firebase": "^6.6.0",
+    "firebase": "^8.2.9",
     "font-awesome": "^4.7.0",
     "jquery": "^3.4.1",
     "ng2-validation": "^4.2.0",
     "popper.js": "^1.15.0",
     "rxjs": "^6.5.3",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.3",
-    "@angular/cli": "8.3.3",
-    "@angular/compiler-cli": "8.2.5",
-    "@angular/language-service": "8.2.5",
+    "@angular-devkit/architect": "0.1102.3",
+    "@angular-devkit/build-angular": "~0.1102.3",
+    "@angular/cli": "11.2.3",
+    "@angular/compiler-cli": "11.2.4",
+    "@angular/language-service": "11.2.4",
     "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^12.7.4",
     "codelyzer": "^5.0.1",
+    "firebase-tools": "^9.5.0",
+    "fuzzy": "^0.1.3",
+    "inquirer": "^6.2.2",
+    "inquirer-autocomplete-prompt": "^1.0.1",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.3.0",
+    "karma": "^6.1.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
     "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "^6.0.0",
+    "protractor": "^7.0.0",
     "ts-node": "~8.3.0",
-    "tslint": "~5.19.0",
-    "typescript": "^3.5.3",
-    "@angular-devkit/architect": "<0.900 || ^0.900.0-0 || ^9.0.0-0",
-    "firebase-tools": "^6.10.0",
-    "fuzzy": "^0.1.3",
-    "inquirer": "^6.2.2",
-    "inquirer-autocomplete-prompt": "^1.0.1"
+    "tslint": "~6.1.3",
+    "typescript": "^4.1.5"
   }
 }
